Replace telescope
We replace the grep functionality from Neovim's telescope plugin with
the native grep functionality provided by Vim. We extend this with the
`ag` (Silver-Surfer) command line tool for better performance.

The functionality to replace files is replaced with a custom function to
utilize Vim's builtin find functionality. We extend this as well with
the `fd` command line utility.