Rewrite logic to grep text and find files
Mostly inspired by [this](https://gist.github.com/romainl/56f0c28ef953ffc157f36cc495947ab3).
When using Vim and grepping stuff the most annoying point surely is that Vim
executes the external grep command in it's parent shell with the following
implications (quoted from the above):

* Vim is suspended for how long it takes the command to finish.
* The results are printed in the shell and clog our terminal's scroll buffer.
* We must press the dreaded <CR> to come back to Vim.

The rewrite of the custom CGrep function fixes this. To enable the same
for finding files we refactored the CFind function in a similar way.

In addition I removed the custom logic to deal with splits and the quickfix
window. I am not sure how satisfied I will be with the current setup but for
now this should suffice.