Replace loclist when looking for files
Working with the Quickfix or Location list makes no sense here as we are
always forced to jump to the first line and first column. We may of course
change the line and column number, but we cannot obtain the previous marked
cursor position without reading and parsing the Shada file. This seems to be
totally overblown so we go with a approach that's similar to that of
Vim-Dirvish, and simply pipe the results of our find command into and
read-only unnamed buffer. Than we can jump to these files via 'gf' or similar
commands.

It would be great if we could simply pipe the obtained files into another
Dirvish buffer to get all of it's nice benefits. But this should suffice for
now.